+++ ipc/helper/binary.go
@@ -14,9 +14,9 @@ func U8aConcat(a []byte, b []byte) []byte {
 
 // U32To8Concat 根据count生成4字节slice，并和b合并
 // 用于格式化数据为：前4字节表示消息大小，后续字节表示消息
-func U32To8Concat(count uint32, b []byte) []byte {
+func U32To8Concat(count int, b []byte) []byte {
 	n := make([]byte, 4)
-	binary.LittleEndian.PutUint32(n, count)
+	binary.LittleEndian.PutUint32(n, uint32(count))
 	return U8aConcat(n, b)
 }
